Movement and Swinging Basics
I added basic movement and animations (should be changed later but not a priority). I also added the grappling mechanic, it has alot of variables that we can change so it can be finetuned when we start making levels. I made the character a ragdoll because the animations were really simple and so was adding the grappling and grabbing mechanics, the characters sprites can be changed.